refactor(remark-html-directives-integration): Fix all file imports to refer to project
