+++ package.json
@@ -12,25 +12,25 @@
     "@fontsource/palanquin": "^5.0.5",
     "@fontsource/raleway": "^5.0.5",
     "@nx/eslint-plugin-nx": "16.0.0-beta.1",
-    "@nx/js": "16.5.5",
-    "@nx/linter": "16.5.5",
-    "@nx/vite": "16.5.5",
-    "@nx/workspace": "16.5.5",
+    "@nx/js": "16.10.0",
+    "@nx/linter": "16.10.0",
+    "@nx/vite": "16.10.0",
+    "@nx/workspace": "16.10.0",
     "@nxtensions/astro": "^3.5.0",
     "@release-it/keep-a-changelog": "^4.0.0",
     "@types/node": "^20.3.1",
-    "@typescript-eslint/eslint-plugin": "^5.60.0",
-    "@typescript-eslint/parser": "^5.60.0",
+    "@typescript-eslint/eslint-plugin": "5.62.0",
+    "@typescript-eslint/parser": "5.62.0",
     "@vitest/coverage-c8": "0.32.2",
     "@vitest/ui": "^0.32.2",
     "astro": "^2.6.5",
     "astro-icon": "^0.8.1",
     "dotenv-cli": "^7.2.1",
-    "eslint": "~8.45.0",
+    "eslint": "8.46.0",
     "eslint-config-prettier": "8.8.0",
     "jsdom": "~22.1.0",
-    "nx": "16.5.5",
-    "nx-cloud": "16.1.1",
+    "nx": "16.10.0",
+    "nx-cloud": "16.4.0",
     "prettier": "^3.0.0",
     "publint": "^0.1.16",
     "release-it": "^16.1.3",
@@ -49,3 +49,4 @@
     "vue": "^3.3.4"
   }
 }
+
