+++ packages/remark-html-directives-integration/src/index.ts
@@ -1,24 +1,25 @@
 import type { AstroIntegration } from 'astro';
 
-import remarkHTMLDirectives from '@forastro/remark-html-directives';
+import remarkHTMLDirectives from './lib';
 import remarkDirective from 'remark-directive';
 
-export default (config:Parameters<typeof remarkHTMLDirectives>[0]):AstroIntegration => {
+export default (config: Parameters<typeof remarkHTMLDirectives>[0]): AstroIntegration => {
     return {
-    name: "@forastro/remark-html-directives-integration",
-    hooks: {
-        "astro:config:setup"({ updateConfig}) {
+        name: "@forastro/remark-html-directives-integration",
+        hooks: {
+            "astro:config:setup"({ updateConfig }) {
 
-          
-            updateConfig({
-                markdown: {
-                    remarkPlugins: [
-                    remarkDirective,
-                    remarkHTMLDirectives(config)
-                  ]
-                }
-            })
 
-        }, 
-    }} ;
+                updateConfig({
+                    markdown: {
+                        remarkPlugins: [
+                            remarkDirective,
+                            remarkHTMLDirectives(config)
+                        ]
+                    }
+                })
+
+            },
+        }
+    };
 }
